Update to Scan.java and Added LinkedList.java
Scan.java:
Modified output to place the output into a file, called temp.txt, for the later parts of the language processor to use.

LinkedList.java:
Customized LinkedList data structure to store an integer, the Token, with a string, the Lexeme. Provides the ability to have a dynamically sized array by means of a function increase the size as well as the ability to access linked list nodes outside of the LinkedList class for information retrieval from the nodes.